requirements.txt
neo4j
google-generativeai
python-dotenv